chore: DBXPORTAL-67 Convert to multi-package repo
- Use private license
- Main package must be set to private
- Temporarely use localhost as registry until ours is set up

Side-effects of multi-package repositories:

As soon as Babel finds a `package.json` it stops searching for a
`.babelrc` file. Therefore Babel needs to be pointed to its
configuration in the CLI (`--config-file` option). The same issue
happens in Jest. Here the only solution is loading the Babel
configuration in JavaScript and adding it to the `transform`
configuration.