fix: Use official DB colors for hover states
Note: In general, we use the color shade 200
for hover states, as we already use the color
shade 100 for the background.
However, for the "warning" state, we deviate
from the official DB color (yellow 200), as
there was already a different color than yellow 100
in place for the background color, probably
because the official bright / dark yellow colors
did not look good when paired together.