fix: Loading secondary Buttons
As secondary Buttons have a transparent background, the loading state
did not cover the text. This caused the text to shine through the
loading animation.

As a fix, the loading state will use the hover color from now on.
Usually the loading state happens right after clicking the Button. In
this situation, a hover effect already made the Button darker.

The stories have been adjusted to use a label on the Button to reflect
the real life.