fix: Optimize Checkbox click
When clicking on the Checkbox the focus ring stays as expected. But
clicking on the label lets the Checkbox loose its focus for a short
moment in time and causes ugly flickering. Even worse, when used within
a Modal component, the Modal gains the focus the same time and flickers
even more.

This might affect Chrome only. Safari and Firefox don’t show the focus
ring on Checkboxes except when the keybord is used.