fix: Vertical alignment
In Chrome the number (digit; eg 1) of a line was not aligned with the
prefix (eg S) when using em units. Though the em units work better in
big sizes, it’s more important to fix the default size, as it’s the only
one we use at the moment.