fix: Small code optimization
Better use `??` as it checks for `null || undefined` instead of falsy values.
In case of `||`, if someone passed the number `0` (a valid placeholder text),
this would evaluate to `' '` instead

We should only add comments where the provide actual value.
A comment 'Inline label' for an `inlineLabel` attribute is not one of those
cases

If anyone supplied a custom `onFocus` prop, this would be overwritten.
Maybe put it after `{...otherProps}` and add a check if an onFocus function
has been provided, then call it and `enforceFocusRingInSafari()` afterwards.